+++ Amiga/Tools/AuDeluxe/AuDeluxe/Views/ContentView.swift
@@ -41,6 +41,8 @@ struct ContentView: View {
                         }
                     }
                 }
+
+                LibraryScanStatusView()
                 
                 if settings.musicFolderURL != nil {
                     if showingTrackerView {